fix(finance-correctness): HR transactional money Float -> exact Numeric(18,2) (S4.6 + S4.9)
S4.6 - payroll totals, payslip pay, compensation base and benefit-plan costs (10
columns) moved from binary Float to NUMERIC(18,2), matching the finance tables -
a paycheck must be exact to the cent. Migration 0053 (Postgres-scoped ALTER;
SQLite builds Numeric from the models via create_all). The Decimal/Float
arithmetic that surfaced was fixed properly: the compensation display band uses
float (a shown estimate); the stored payslip gross stays exact Decimal and the
run accumulators start at Decimal("0").

The ~17 other money-NAMED Float columns (org-graph budgets, ARR/MRR, per-1k-token
cost RATES, statistical estimates) are deliberately left Float - they are
analytics/rates, not transactional amounts, and (18,2) would truncate the token
rates. Refutes the plan's blanket "convert 30 columns".

S4.9 - SorObject.deleted was Mapped[bool] over an Integer column the actuator
reads/writes as 0/1; annotation aligned to Mapped[int] (no migration).

Verified: drift gate green on pg16 (256 tables, 0 unaccepted diffs); ruff clean;
HR full-coverage + api-gated + fairness 30 passed; finance_seed 2 passed;
tax_agent reads these read-only, no non-HR arithmetic consumers.

+++ backend/app/hr/models/benefits.py
@@ -2,7 +2,7 @@
 KAEOS HR Vertical — Benefits Models
 Function 4: Benefits Administration
 """
-from sqlalchemy import Column, String, Float, DateTime, Boolean, JSON, ForeignKey, Enum, Date
+from sqlalchemy import Column, String, Numeric, DateTime, Boolean, JSON, ForeignKey, Enum, Date
 from sqlalchemy.sql import func
 import enum
 
@@ -34,9 +34,9 @@ class BenefitPlan(Base):
     coverage_details = Column(JSON, default=dict)
     
     # Costs (Monthly)
-    employee_cost_individual = Column(Float, default=0.0)
-    employee_cost_family = Column(Float, default=0.0)
-    employer_contribution = Column(Float, default=0.0)
+    employee_cost_individual = Column(Numeric(18, 2), default=0)
+    employee_cost_family = Column(Numeric(18, 2), default=0)
+    employer_contribution = Column(Numeric(18, 2), default=0)
     
     is_active = Column(Boolean, default=True)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
